cyclictest: fix MODE_CYCLIC measurement failure after timer overruns
This patch aims to fix nonsensical results sometimes reported by
cyclictest when running in MODE_CYCLIC.

In MODE_CYCLIC if we hit the "now > next" case, we can end up adjusting
the "next" timespec twice: by the tsgreater-while-loop, and after that
on a later iteration of the loop by the result of timer_getoverrun().

After this double adjustment, the results will not be correct.

Signed-off-by: Tommi Rantala <tommi.t.rantala@nokia.com>